Restructure cmrc resource lib compiling to allow linking genkiml to multiple targets
