+++ examples/keras/fc_model/fingerprint.pb
