+++ examples/keras/fc_model/keras_metadata.pb
