+++ examples/keras/fc_model/saved_model.pb
