+++ examples/keras/fc_model/variables/variables.index
