+++ genkiml/source/genkiml.cpp
@@ -6,10 +6,7 @@
 #include <range/v3/numeric.hpp>
 #include <string>
 
-#include "cmrc/cmrc.hpp"
 #include "genkiml.h"
-
-CMRC_DECLARE(files);
 
 namespace genki::ml {
 
@@ -206,24 +203,5 @@ Model::Model(gsl::span<const gsl::byte> model_data) : impl(std::make_unique<Impl
 auto Model::infer(const BufferViews& input) -> BufferViews { return impl->infer(input); }
 
 Model::~Model() = default;
-
-//======================================================================================================================
-std::unique_ptr<Model> load_model()
-{
-    constexpr auto prefix = "models";
-
-    auto fs         = cmrc::files::get_filesystem();
-    auto models_dir = fs.iterate_directory(prefix);
-    assert(std::distance(models_dir.begin(), models_dir.end()) == 1); // Only embedding one model atm
-
-    const auto model = fmt::format("{}/{}", prefix, (*models_dir.begin()).filename());
-
-    assert(fs.is_file(model));
-
-    const auto model_file = fs.open(model);
-    const auto bytes      = gsl::as_bytes(gsl::span(model_file.cbegin(), model_file.cend()));
-
-    return std::make_unique<Model>(bytes);
-}
 
 } // namespace genki::ml
